requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai>=1.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
streamlit>=1.36.0
